requirements.txt
streamlit
pandas
numpy
scikit-learn
langgraph
langchain
langchain-groq
python-dotenv